Refactor away the api namespace into a module
